Tests/LibWeb/Text/expected/HTML/ValidityState-valueMissing.txt

text (required) - Initial valueMissing: true
text (required) - After setting value - valueMissing: false
text (required) - After clearing value - valueMissing: true
text (optional) - Initial valueMissing: false
text (optional) - After setting value - valueMissing: false
text (optional) - After clearing value - valueMissing: false
checkbox (required) - Initial valueMissing: true
checkbox (required) - After setting value - valueMissing: false
checkbox (required) - After clearing value - valueMissing: true
radio (required) - Initial valueMissing: true
radio (required) - After setting value - valueMissing: false
radio (required) - After clearing value - valueMissing: true
number (required) - Initial valueMissing: true
number (required) - After setting value - valueMissing: true
number (required) - After clearing value - valueMissing: true
date (required) - Initial valueMissing: true
date (required) - After setting value - valueMissing: true
date (required) - After clearing value - valueMissing: true
textOptional - valueMissing when empty: false
text (required) - Disabled - valueMissing: false
text (optional) - Disabled - valueMissing: false
checkbox (required) - Disabled - valueMissing: true
radio (required) - Disabled - valueMissing: true
number (required) - Disabled - valueMissing: false
date (required) - Disabled - valueMissing: false
